src/main/java/org/generation/blogpessoal/repository/UsuarioRepository.java
package org.generation.blogpessoal.repository;

import java.util.List;
import java.util.Optional;

import org.generation.blogpessoal.model.Usuario;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

@Repository
public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
	public Optional<Usuario> findByUsuario(String usuario);
	
	public List<Usuario> findAllByNomeContainingIgnoreCase(String nome);
	// teste para find por nome*

}

//Optional concede a possibilidade de n respostas possiveis na conferencia do usu na bd... achar qualquer resposta contida

